Refactor libcheckisomd5
* Use of page-aligned buffer.
* Advertise c11 to the compiler.
* Move common dependencies of libcheckisomd5 and libimplantisomd5 to
shared utilities header.
* Avoid using magic values.
* Declare right before use and trust the compiler with optimizations.
* Make more use of size_t and use off_t where 64 bit me needed.
* Try to make it more compatible with Windows.
* Simplify code to clean appdata.
* Move important define into utilities.

+++ pyisomd5sum.c
@@ -18,6 +18,7 @@
 
 #include <Python.h>
 #include <stdio.h>
+#include <sys/types.h>
 
 #include "libcheckisomd5.h"
 #include "libimplantisomd5.h"
@@ -34,7 +35,7 @@ static PyMethodDef isomd5sumMethods[] = {
 /* Call python object with offset and total
  * If the object returns true return 1 to abort the check
  */
-int pythonCB(void *cbdata, long long offset, long long total) {
+int pythonCB(void *cbdata, off_t offset, off_t total) {
     PyObject *arglist, *result;
     int rc;
